i guess i forgot to add the loadout image. I'll change that image later on. It's kind of a placeholder atm.
I guess it wont run without the image file. I added a sidebar to the text editor that adds classes and methods to a listbox. It makes it really easy to jump to different parts of your code. I also created call backs to change the  color of functions and methods. I plan on going back through and using regular expressions to make the callback better at discerning when a class or method was used. They keep picking up variables I used while writing the code lol.